18.04.2016

+++ src/main/java/dao/EmployeeDao.java
@@ -8,12 +8,12 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
 @Transactional
-public interface EmployeeDao extends CrudRepository<Employee, Long> {
+public interface EmployeeDao extends CrudRepository<Employee, Integer> {
 
   public Employee findByUserId(String userId);
   
   public Employee findById(int id);
     
   @Query("SELECT new model.Employees(p.id, p.firstName, p.lastName, p.organizationName, p.subOrganizationName, p.mobilephoneNumber, p.telephoneNumber, p.smallImage) FROM Employee p WHERE p.firstName like %:term% OR p.lastName LIKE %:term% OR p.organizationName LIKE %:term% OR p.subOrganizationName LIKE %:term% OR p.jobTitle LIKE %:term% OR p.mobilephoneNumber LIKE %:term% OR p.telephoneNumber LIKE %:term% OR p.userId LIKE %:term% OR p.location LIKE %:term% ORDER by p.lastName")
-  public Iterable<Employee> findEmployeesBySearchTerm(@Param("term") String term);  
+  public Iterable<Employee> findEmployeesBySearchTerm(@Param("term") String term);
 }
